.requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migration
psycopg2-binary
